+++ tests/__init__.py
@@ -2,3 +2,6 @@ import os
 
 # turn off caching
 os.environ["NOCACHE"] = "True"
+
+# turn off rate limits
+os.environ["NORATELIMIT"] = "True"