+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 Notable changes to this project.
 
+## [dev]
+- rate limit requests to the numerai api
+- allow turning of rate limiting
+- test: increase coverage
+
 ## [0.2.3] - 2019-06-27
 - added `friends` report - showing correlation of a user's live auroc to other users
 - fix `payments` report for rounds >= 158